% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectScerevisiae.R
\docType{class}
\name{TrenaProjectScerevisiae-class}
\alias{TrenaProjectScerevisiae-class}
\alias{.TrenaProjectScerevisiae}
\alias{TrenaProjectScerevisiae}
\title{TrenaProjectScerevisiae-class}
\usage{
TrenaProjectScerevisiae(quiet = TRUE)
}
\value{
An object of the TrenaProjectScerevisiae class
}
\description{
Expression, variant and covariate data for the genes of interest (perhaps unbounded) for pre-term birth studies
}
